requirements.txt
boto3==1.9.96
Flask==1.0.2
hurry.filesize==0.9
psutil==5.5.1
requests==2.21.0
twarc==1.6.1
python-dateutil==2.8.0
honeybadger==0.2.0
